recycle: move task to core
