+++ caikit/core/data_model/__init__.py
@@ -29,4 +29,3 @@ from .enums import *
 from .producer import PACKAGE_COMMON, ProducerId
 from .streams import data_stream
 from .streams.data_stream import *
-from .task_decorator import task
